fix modules classes
