+++ modules/Books-dataClass.js
@@ -1,0 +1,7 @@
+export default class BooksDataClass {
+  constructor() {
+    return [];
+  }
+}
+
+export const booksData = new BooksDataClass();
