+++ modules/validations.js
@@ -1,5 +1,3 @@
-/* eslint-disable max-classes-per-file */
-
 const error = document.querySelector('#error');
 
 const showmessage = (test, msg) => {
@@ -11,11 +9,13 @@ const showmessage = (test, msg) => {
   return test;
 };
 
-export default function validations(input, msg, input2, msg2) {
+const validations = (input, msg, input2, msg2) => {
   if (input.trim() === '') {
     return showmessage(false, msg);
   } if (input2.trim() === '') {
     return showmessage(false, msg2);
   }
   return showmessage(true, '');
-}
+};
+
+export default validations;
